+++ tests/tests.py
@@ -383,6 +383,10 @@ def test_fasta():
 	run('-a TTAGACATATCTCCGTCG', 'small.fasta', 'small.fastq')
 
 
+def test_fasta_no_trim():
+	run([], 'small-no-trim.fasta', 'small.fastq')
+
+
 def test_issue_202():
 	"""Ensure --length-tag= also modifies the second header line"""
 	run('-a GGCTTC --length-tag=length=', 'SRR2040271_1.fastq', 'SRR2040271_1.fastq')